requirements.txt
flowserv-core>=0.7.1
flask
flask_cors
